Added vessel steering and moving methods

+++ src/submarine/Submarine.java
@@ -8,7 +8,7 @@ import commons.Vessel;
 
 public class Submarine extends Vessel {
 	
-	private float depth;
-	private float maxDepth;
+	private double depth;
+	private double maxDepth;
 
 }
